Add echo statements

+++ post-install-user.sh
@@ -1,12 +1,23 @@
 #!/usr/bin/env bash
 
-set -euxo pipefail
+set -euo pipefail
+
+echo "** Starting userspace post install script"
 
 # User SSH setting
-mkdir $HOME/.ssh
-curl -SsL https://github.com/narbux.keys -o .ssh/authorized_keys
+echo -e "\t>> Downloading ssh keys"
+if [[ ! -d "$HOME"/.ssh ]]; then
+    mkdir $HOME/.ssh
+fi
+
+if [[ -e /usr/bin/curl ]]; then
+    curl -SsL https://github.com/narbux.keys -o .ssh/authorized_keys
+else
+    echo -e "\t>>\033[31m\033ERROR:[0m Could not find curl to download SSH keys"
+fi
 
 # Install PARU AUR helper
+echo -e "\t>> Downloading Paru"
 git clone --depth=1 https://aur.archlinux.org/paru-bin \
     && cd paru-bin \
     && makepkg -si --noconfirm \
@@ -14,6 +25,7 @@ git clone --depth=1 https://aur.archlinux.org/paru-bin \
     && rm -rf paru-bin
 
 # Install and configure zsh-antidote and zsh
+echo -e "\t>> Downloading ZSH-antidote and configuring ZSH"
 paru -S --noconfirm zsh-antidote
 
 cat <<'EOF' >> $HOME/.zsh_plugins.txt
@@ -50,4 +62,7 @@ autoload -Uz promptinit && promptinit && prompt pure
 eval "$(zoxide init zsh)"
 EOF
 
+echo -e "\t>> Removing Bash leftover files"
 rm $HOME/.bash_logout $HOME/.bash_profile $HOME/.bashrc
+
+echo "** DONE **"
